+++ program/tests/helpers/test-transfer-hook/Cargo.toml
@@ -2,6 +2,7 @@
 name = "test-transfer-hook"
 version = "0.1.0"
 edition = "2021"
+publish = false
 
 [features]
 no-entrypoint = []
